--- a/packages/cache/src/internal/constants.ts
+++ b/packages/cache/src/internal/constants.ts
@@ -5,6 +5,7 @@ export enum CacheFilename {
 
 export enum CompressionMethod {
   Gzip = 'gzip',
+  Pigz = 'pigz',
   // Long range mode was added to zstd in v1.3.2.
   // This enum is for earlier version of zstd that does not have --long support
   ZstdWithoutLong = 'zstd-without-long',