Adds support for `pigz` as a compression method.

`pigz` stands for "**P**arallel **I**mplementation of **Gz**ip". As the name suggests, it tries to do the work in parallel and so it's usually faster than plain Gzip.

Right now it'll just pick `pigz` over `gzip` when it's available. Usually you'll need to manually install `pigz` on the system and then this action can pick that up.

I think I'll need to write some tests? Can someone point out how?
Considering the implementation relies on gzip being installed before-hand.

#### I'm not a js/ts developer please point out any sub-optimalities. Not to mention first contribution to this repo.

## Why though?

For the Swift Package Manager (SwiftPM for short), we cache the `.build` directory for faster build times.
This helps a ton and can make things easily 2.5x faster. The bigger the project, the bigger the impact.
Swift also has a pretty slow build system, so the difference between build times on a real project is at least 3 mins. I've seen 20+ minutes of saved build time too.

The problem is, SwiftPM's `.build` directory is relatively big. Starts from like 500 MB for a non-insignificant project, and can easily climb up to 2 GB+.
The cache action needs to compress / decompress all this data, and that means it can take a few minutes just for caching stuff. The decompression is usually fast enough though, not reaching a minute even for 2 GB+ files.

## Numbers?

I've done some tests manually on some projects.

| Cache Size (GB) | Save Cache Time | Time Ratio | Machine Size     | Project         |
|-----------------|-----------------|------------|------------------|-----------------|
| 0.6             | 57s -> 46s      | 1.24x        | GitHub's Default | Private 1       |
| 1               | [76s](https://github.com/vapor/penny-bot/actions/runs/12547587263/job/34985338526) -> [37s](https://github.com/vapor/penny-bot/actions/runs/12557099546/job/35009211708)      | 2.05x        | GitHub's Default | vapor/penny-bot |
| 1.8             | 153s -> 43s     | 3.56x       | EC2 32 core      | Private 2       |
| 2.1             | 210s -> 53s     | 3.96x       | EC2 32 core      | Private 2       |