@MahdiBM 2025-01-01
Nevermind, just found out about zstd and it's even faster than pigz for .build caching ...
Suboptimal that this trick isn't documented anywhere.